✨ New candidacy flow (#174)
* ✨ Staking struct

* ♻️  Rename Consensus::validators to pubkeys

* ✨ New candidacy flow

* ♻️  Use bonded amount as voting power & refacto crypto to self

* clippy

* genesis bond

* 🐛 Fix network read missing messages

* add catchup request for staking info & add endpoint to set stake

* use expect

* add pubkey to registry on new bond

* ✅ Fix test witch catchup

* 🔊 do not remove log

* remove constants.Rs

* minor changes after review

* 🐛 Bond new validators

* 💡 Add comments

+++ src/consensus/metrics.rs
@@ -99,11 +99,11 @@ impl ConsensusMetrics {
         self.confirm_ack_commit_aggregate.add(1, &[]);
     }
 
-    pub fn confirmed_ack_gauge(&self, nb: usize) {
-        self.confirmed_ack_gauge.record(nb as u64, &[])
+    pub fn confirmed_ack_gauge(&self, nb: u64) {
+        self.confirmed_ack_gauge.record(nb, &[])
     }
-    pub fn prepare_votes_gauge(&self, nb: usize) {
-        self.prepare_votes_gauge.record(nb as u64, &[])
+    pub fn prepare_votes_gauge(&self, nb: u64) {
+        self.prepare_votes_gauge.record(nb, &[])
     }
 
     pub fn commit(&self) {
